💄 #15 style: 모달 반응형 디자인 수정 외
- 텍스트에디터 서식모음 컴포넌트 상위로 이동
