+++ app/(nav)/todos/[todoId]/_view/NoteForm.tsx
@@ -24,6 +24,18 @@ import IconCheck from '@/public/icons/IconCheck';
 import { useCurrentEditor } from '@tiptap/react';
 import LinkEmbed from './LinkEmbed';
 import IconModalClose from '@/public/icons/IconModalClose';
+import IconTextAlignLeft from '@/public/icons/IconTextAlignLeft';
+import IconTextAlignMiddle from '@/public/icons/IconTextAlignMiddle';
+import IconTextAlignRight from '@/public/icons/IconTextAlignRight';
+import IconTextBold from '@/public/icons/IconTextBold';
+import IconTextBulletPoint from '@/public/icons/IconTextBulletPoint';
+import IconTextColor from '@/public/icons/IconTextColor';
+import IconTextItalics from '@/public/icons/IconTextItalics';
+import IconTextNumberPoint from '@/public/icons/IconTextNumberPoint';
+import IconTextUnderline from '@/public/icons/IconTextUnderline';
+import IconTextHighlight from '@/public/icons/IconTextHighlight';
+import IconAddLink from '@/public/icons/IconAddLink';
+import InputSlid from '@/components/common/InputSlid';
 
 type NoteFormProps = {
   title?: string;
@@ -49,7 +61,7 @@ const NoteForm = ({ title: initTitle = '', linkUrl: initLinkUrl = '', method = '
   const handleChangeTitle: ChangeEventHandler<HTMLInputElement> = (e) => {
     setTitle(e.target.value.length > 30 ? e.target.value.slice(0, 30) : e.target.value);
   };
-  const handleSaveLinkUrl = (linkUrlValue: string) => setLinkUrl(linkUrlValue);
+  const handleSaveLinkUrl = () => setLinkUrl(linkUrlValue);
 
   const handleSave = useCallback(() => {
     window.localStorage.setItem(
@@ -120,6 +132,25 @@ const NoteForm = ({ title: initTitle = '', linkUrl: initLinkUrl = '', method = '
     });
   };
 
+  const [linkUrlValue, setLinkUrlValue] = useState(linkUrl);
+
+  const handleChangeLinkUrlValue: ChangeEventHandler<HTMLInputElement | HTMLSelectElement> = (e) =>
+    setLinkUrlValue(e.target.value);
+
+  const handleBold = () => editor?.chain().focus().toggleBold().run();
+  const handleItalics = () => editor?.chain().focus().toggleItalic().run();
+  const handleUnderline = () => editor?.chain().focus().toggleUnderline().run();
+  const handleAlignLeft = () => editor?.chain().focus().setTextAlign('left').run();
+  const handleAlignCenter = () => editor?.chain().focus().setTextAlign('center').run();
+  const handleAlignRight = () => editor?.chain().focus().setTextAlign('right').run();
+  const handleBulletList = () => editor?.chain().focus().toggleBulletList().run();
+  const handleOrderedList = () => editor?.chain().focus().toggleOrderedList().run();
+  const handleChangeHighlight: ChangeEventHandler<HTMLInputElement> = (e) =>
+    editor?.chain().focus().toggleHighlight({ color: e.target.value }).run();
+
+  const handleChangeColor: ChangeEventHandler<HTMLInputElement> = (e) =>
+    editor?.chain().focus().setColor(e.target.value).run();
+
   return (
     <>
       {linkUrl && isEmbedOpen && (
@@ -134,7 +165,7 @@ const NoteForm = ({ title: initTitle = '', linkUrl: initLinkUrl = '', method = '
           </div>
         </section>
       )}
-      <section className='lg:max-w-[800px] w-full p-4 md:p-6 lg:py-6 lg:px-10 flex flex-col grow'>
+      <section className='lg:max-w-[800px] w-full p-4 md:p-6 lg:py-6 lg:px-10 flex flex-col grow '>
         <div className='flex w-full items-center mb-4'>
           <h1 className='grow text-slate-900 font-semibold text-lg'>노트 작성</h1>
           <button className='py-3 px-5 text-blue-500 font-semibold text-sm mr-2' onClick={handleSave}>
@@ -189,7 +220,7 @@ const NoteForm = ({ title: initTitle = '', linkUrl: initLinkUrl = '', method = '
           </div>
         )}
         <hr />
-        <form className='grow w-full h-fit relative flex flex-col' onSubmit={handleSubmit}>
+        <form className='grow w-full relative flex flex-col' onSubmit={handleSubmit}>
           <div className='w-full relative h-7 my-3'>
             <input
               className='w-full text-lg font-medium focus-visible:outline-none'
@@ -217,11 +248,80 @@ const NoteForm = ({ title: initTitle = '', linkUrl: initLinkUrl = '', method = '
               </div>
             </div>
           )}
-          <div className='grow'>
-            <TiptapEditor linkUrl={initLinkUrl} onSaveLinkUrl={handleSaveLinkUrl} />
+          <div className='grow lg:relative'>
+            <div className='overflow-visible h-full'>
+              <TiptapEditor linkUrl={initLinkUrl} onSaveLinkUrl={handleSaveLinkUrl} />
+            </div>
           </div>
+          <div className='lg:max-w-[768px] border border-slate-200 rounded-full py-2.5 px-4 fixed lg:sticky bottom-4 left-4 right-4 bg-white flex gap-2 md:gap-4 overflow-auto scrollbar-width-none'>
+            <div className='flex md:gap-1'>
+              <button onClick={handleBold}>
+                <IconTextBold className='cursor-pointer hover:bg-slate-100' />
+              </button>
+              <button onClick={handleItalics}>
+                <IconTextItalics className='cursor-pointer hover:bg-slate-100' />
+              </button>
+              <button onClick={handleUnderline}>
+                <IconTextUnderline className='cursor- hover:bg-slate-100' />
+              </button>
+            </div>
+            <div className='flex md:gap-1'>
+              <button onClick={handleAlignLeft}>
+                <IconTextAlignLeft className='cursor-pointer hover:bg-slate-100' />
+              </button>
+              <button onClick={handleAlignCenter}>
+                <IconTextAlignMiddle className='cursor-pointer hover:bg-slate-100' />
+              </button>
+              <button onClick={handleAlignRight}>
+                <IconTextAlignRight className='cursor-pointer hover:bg-slate-100' />
+              </button>
+            </div>
+            <div className='flex md:gap-1'>
+              <button onClick={handleBulletList}>
+                <IconTextBulletPoint className='cursor-pointer hover:bg-slate-100' />
+              </button>
+              <button onClick={handleOrderedList}>
+                <IconTextNumberPoint className='cursor-pointer hover:bg-slate-100' />
+              </button>
+              <label htmlFor='textHighlight' className='relative'>
+                <input id='textHighlight' type='color' className='w-0 h-0 absolute' onChange={handleChangeHighlight} />
+                <IconTextHighlight className='cursor-pointer hover:bg-slate-100' />
+              </label>
+              <label htmlFor='textColor' className='relative'>
+                <input id='textColor' type='color' className='w-0 h-0 absolute' onChange={handleChangeColor} />
+                <IconTextColor className='cursor-pointer hover:bg-slate-100' />
+              </label>
+            </div>
+            <div className='grow flex justify-end'>
+              <ModalProvider>
+                <ModalTrigger type='button'>
+                  <IconAddLink className='cursor-pointer hover:bg-slate-100' />
+                </ModalTrigger>
+                <ModalContent className='w-full max-w-[520px] flex flex-col'>
+                  <div className='flex justify-between mb-6'>
+                    <h1 className='text-lg font-bold'>링크 업로드</h1>
+                    <ModalClose />
+                  </div>
+                  <InputSlid
+                    label='링크'
+                    type='text'
+                    placeholder='영상이나 글, 파일의 링크를 넣어주세요'
+                    className='mb-10'
+                    value={linkUrlValue}
+                    onChange={handleChangeLinkUrlValue}
+                  />
+                  <ModalClose asChild>
+                    <Button className='w-full' onClick={handleSaveLinkUrl}>
+                      확인
+                    </Button>
+                  </ModalClose>
+                </ModalContent>
+              </ModalProvider>
+            </div>
+          </div>
+
           {savedToast && (
-            <div className='max-w-[768px] fixed bottom-0 left-4 right-4 -translate-y-[155%] bg-blue-50 text-blue-500 rounded-full py-2.5 px-6 flex gap-2 items-center'>
+            <div className='max-w-[768px] fixed lg:sticky bottom-0 left-4 right-4 -translate-y-[155%] bg-blue-50 text-blue-500 rounded-full py-2.5 px-6 flex gap-2 items-center'>
               <IconCheck />
               <p className='font-semibold text-sm'>
                 임시 저장이 완료되었습니다 <span className='text-xs pointerfont-medium'>ㆍ {}초전</span>
