+++ app/(nav)/todos/[todoId]/create/page.tsx
@@ -3,7 +3,7 @@ import NoteForm from '../_view/NoteForm';
 
 export default function Page() {
   return (
-    <main className='lg:flex h-screen lg:flex-row w-screen lg:w-full'>
+    <main className='lg:flex h-screen w-screen'>
       <TiptapEditorProvider
         className='resize-none w-full h-full focus-visible:outline-none text-slate-700 whitespace-break-spaces'
         slotBefore={<NoteForm />}
