+++ app/(nav)/todos/[todoId]/note/[noteId]/page.tsx
@@ -16,7 +16,7 @@ export default async function Page({ params }: { params: { noteId: string } }) {
   return (
     <main className='lg:flex h-screen w-screen'>
       <TiptapEditorProvider
-        className='resize-none w-full h-full focus-visible:outline-none text-slate-700 whitespace-break-spaces'
+        className='resize-none w-full h-full focus-visible:outline-none text-slate-700 whitespace-break-spaces '
         content={content}
         slotBefore={<NoteForm title={title} content={content} linkUrl={linkUrl} method='PATCH' noteId={noteId} />}
       />
