+++ components/common/Modal.tsx
@@ -68,7 +68,7 @@ const ModalContent = ({ className, children, ...props }: ComponentPropsWithoutRe
     <>
       {isOpen &&
         createPortal(
-          <div className='fixed inset-0 flex justify-center items-center' onClick={handleClickOverlay}>
+          <div className='fixed inset-0 flex justify-center items-center p-8' onClick={handleClickOverlay}>
             <div className='absolute inset-0 bg-black opacity-50'></div>
             <div className={twMerge('p-6 rounded-xl bg-white z-10', className)} ref={ref} {...props}>
               {children}
